reactive-review: move users to user module
