+++ frontend/src/store/modules/review/actions.js
@@ -10,7 +10,7 @@ export default {
                 description: review.description
             })
                 .then(function (res) {
-                    context.commit('ADD_USER', {
+                    context.commit('userList/ADD_REVIEW_USER', {
                         data: context.rootState.auth.currentUser
                     });
                     const review = Object.assign({}, res.data.data);
