+++ frontend/src/store/modules/review/mutations.js
@@ -2,23 +2,6 @@ import normalizer from '../../../services/common/normalizerService';
 import { STATUS_NONE } from '@/services/api/codes';
 
 export default {
-    ADD_USER: (state, user) => {
-        if (state.users.allIds.includes(user.data.id)) {
-            return;
-        }
-        const newUser = normalizer.normalize(user, {
-            first_name: '',
-            last_name: '',
-            avatar_url: ''
-        });
-        state.users.byId = Object.assign(
-            {},
-            state.users.byId,
-            newUser.byId
-        );
-        state.users.allIds.push(user.data.id);
-    },
-
     ADD_REVIEW: (state, review) => {
         const newReview = normalizer.normalize( { data: review }, state.getReviewSchema());
         state.reviews.byId = Object.assign(
@@ -31,10 +14,6 @@ export default {
 
     SET_CURRENT_PLACE_REVIEWS: (state, reviews) => {
         state.reviews = reviews;
-    },
-
-    SET_CURRENT_PLACE_REVIEWS_USERS: (state, users) => {
-        state.users = users;
     },
 
     SET_CURRENT_PLACE_REVIEW_LIKE_STATE: (state, { reviewId, likeState }) => {
