+++ frontend/src/store/modules/userList/state.js
@@ -1,3 +1,4 @@
 export default {
+    normalizeReviewUsers: null,
     userLists : null,
 };
